relocated gpu memory
